+++ eth_defi/lagoon/deployment.py
@@ -30,7 +30,7 @@ from web3.contract import Contract
 from web3.contract.contract import ContractFunction
 
 from eth_defi.aave_v3.deployment import AaveV3Deployment
-from eth_defi.abi import get_deployed_contract
+from eth_defi.abi import get_deployed_contract, ZERO_ADDRESS_STR
 from eth_defi.deploy import deploy_contract
 from eth_defi.erc_4626.vault import ERC4626Vault
 from eth_defi.foundry.forge import deploy_contract_with_forge
@@ -404,11 +404,13 @@ def deploy_lagoon(
     factory_contract=True,
     beacon_address="0x652716FaD571f04D26a3c8fFd9E593F17123Ab20",
     beacon_proxy_factory_address=None,
+    beacon_proxy_factory_abi="lagoon/BeaconProxyFactory.json",
     vault_abi="lagoon/v0.5.0/Vault.json",
     deploy_fee_registry: bool = True,
     fee_registry_address: HexAddress | None = None,
     legacy: bool = False,
     salt=Web3.to_bytes(hexstr="0x" + "01" * 32),
+    optin_proxy_delay=3 * 24 * 3600,
 ) -> Contract:
     """Deploy a new Lagoon vault.
 
@@ -554,18 +556,41 @@ def deploy_lagoon(
         assert beacon_proxy_factory_address, f"Cannot deploy Lagoon vault beacon proxy on chain {chain_id}, no factory address found. Registered factories: {pformat(LAGOON_BEACON_PROXY_FACTORIES)}"
         beacon_proxy_factory = get_deployed_contract(
             web3,
-            "lagoon/BeaconProxyFactory.json",
+            beacon_proxy_factory_abi,
             beacon_proxy_factory_address,
         )
 
-        args = [parameters.get_create_vault_proxy_arguments(), salt]
-        logger.info(
-            "Transacting with factory contract %s.createVaultProxy() with args %s",
-            beacon_proxy_factory_address,
-            args,
-        )
-
-        bound_func = beacon_proxy_factory.functions.createVaultProxy(*args)
+        # Deal with unstable ABI madness
+        match beacon_proxy_factory_abi:
+            case "lagoon/BeaconProxyFactory.json":
+                # Leacy
+                args = [parameters.get_create_vault_proxy_arguments(), salt]
+                logger.info(
+                    "Transacting with factory contract %s.createVaultProxy() with args %s",
+                    beacon_proxy_factory_address,
+                    args,
+                )
+                bound_func = beacon_proxy_factory.functions.createVaultProxy(*args)
+            case "lagoon/OptinProxyFactory.json":
+                # https://docs.lagoon.finance/vault/create-your-vault
+                assert len(salt) == 32
+                args = [
+                    ZERO_ADDRESS_STR,  # _logic
+                    safe.address,  # __initialOwner
+                    optin_proxy_delay,  # _initialDelay
+                    parameters.get_create_vault_proxy_arguments(),
+                    salt,
+                ]
+                logger.info(
+                    "Transacting with OptinBeaconFactory contract %s.createVaultProxy() with args %s",
+                    beacon_proxy_factory_address,
+                    args,
+                )
+                bound_func = beacon_proxy_factory.functions.createVaultProxy(
+                    *args,
+                )
+            case _:
+                raise NotImplementedError(f"Unknown Lagoon proxy factory ABI pattern: {beacon_proxy_factory_abi}")
 
         tx_params = {
             "gas": 2_000_000,
@@ -579,7 +604,13 @@ def deploy_lagoon(
         assert_transaction_success_with_explanation(web3, tx_hash)
 
         receipt = web3.eth.get_transaction_receipt(tx_hash)
-        events = beacon_proxy_factory.events.BeaconProxyDeployed().process_receipt(receipt)
+        match beacon_proxy_factory_abi:
+            case "lagoon/BeaconProxyFactory.json":
+                events = beacon_proxy_factory.events.BeaconProxyDeployed().process_receipt(receipt)
+            case "lagoon/OptinProxyFactory.json":
+                events = beacon_proxy_factory.events.ProxyDeployed().process_receipt(receipt)
+            case _:
+                raise NotImplementedError(f"Unknown Lagoon proxy factory ABI pattern: {beacon_proxy_factory_abi}")
         event = events[0]
         contract_address = event["args"]["proxy"]
         vault = get_deployed_contract(
@@ -940,6 +971,7 @@ def deploy_automated_lagoon_vault(
         logger.info("Between contracts deployment delay: Sleeping %s for new nonce to propagade", between_contracts_delay_seconds)
         time.sleep(between_contracts_delay_seconds)
 
+    beacon_proxy_factory_abi = "lagoon/BeaconProxyFactory.json"  # Default ABI (legacy)
     if not existing_vault_address:
         if from_the_scratch:
             # Deploy the full Lagoon protocol with fee registry and beacon proxy factory,
@@ -954,7 +986,10 @@ def deploy_automated_lagoon_vault(
             )
             beacon_proxy_factory_address = beacon_proxy_factory_contract.address
         else:
-            beacon_proxy_factory_address = LAGOON_BEACON_PROXY_FACTORIES.get(chain_id)
+            beacon_factory = LAGOON_BEACON_PROXY_FACTORIES.get(chain_id)
+            assert beacon_factory, f"No beacon factory in LAGOON_BEACON_PROXY_FACTORIES for {chain_id}"
+            beacon_proxy_factory_address = beacon_factory["address"]
+            beacon_proxy_factory_abi = beacon_factory["abi"]
 
         assert beacon_proxy_factory_address, f"Cannot deploy Lagoon vault beacon proxy on chain {chain_id}, no factory address found. Registered factories: {pformat(LAGOON_BEACON_PROXY_FACTORIES)}"
 
@@ -971,6 +1006,7 @@ def deploy_automated_lagoon_vault(
             factory_contract=factory_contract,
             legacy=legacy,
             beacon_proxy_factory_address=beacon_proxy_factory_address,
+            beacon_proxy_factory_abi=beacon_proxy_factory_abi,
         )
 
     if not is_anvil(web3):
@@ -1099,7 +1135,17 @@ LAGOON_FEE_REGISTRIES = {
 #: https://docs.lagoon.finance/vault/create-your-vault
 LAGOON_BEACON_PROXY_FACTORIES = {
     # Base
-    8453: "0xC953Fd298FdfA8Ed0D38ee73772D3e21Bf19c61b",
+    8453: {
+        "abi": "lagoon/BeaconProxyFactory.json",
+        "address": "0xC953Fd298FdfA8Ed0D38ee73772D3e21Bf19c61b",
+    },
     # Arbitrum
-    42161: "0x9De724B0efEe0FbA07FE21a16B9Bf9bBb5204Fb4",
+    # 42161: "0x9De724B0efEe0FbA07FE21a16B9Bf9bBb5204Fb4",
+    # Arbitrum new
+    # Impl https://arbiscan.io/address/0xbb2de8e36eb36dbc20d71c503711763a4be3b1b2#readContract
+    # Proxy https://arbiscan.io/address/0xb1ee4f77a1691696a737ab9852e389cf4cb1f1f5#writeProxyContract#F1
+    42161: {
+        "abi": "lagoon/OptinProxyFactory.json",
+        "address": "0xb1ee4f77a1691696a737ab9852e389cf4cb1f1f5",
+    },
 }